requirements.txt
streamlit
langchain
huggingface-hub
python-dotenv
